+++ tableaudocumentapi/__init__.py
@@ -5,9 +5,11 @@ from .relation import Relation
 from .db_column import DBColumn
 from .connection import Connection
 from .parameter import Parameter
+from .refresh_event import RefreshEvent
+from .refresh import Refresh
 from .extract import Extract
 from .datasource import Datasource, ConnectionParser, ParameterParser
 from .workbook import Workbook
 
-__version__ = '0.9.0'
+__version__ = '0.10.0'
 __VERSION__ = __version__